updates LabelWrapper story - remove 'inValid` control - renames directory to `LabelWrapper`
